+++ HISTORY.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+`NEW` add ButtonGroupFormField used in Grid sub component.
+`NEW` add prop `passedData` to receive values from grid.
+
 ## 1.0.2
 
 `NEW` add common Field & rewrite data flow